Make component/Lobby a stateless component
